+++ backend/models/College.js
@@ -77,7 +77,8 @@ const collegeSchema = new mongoose.Schema({
   verificationToken: String,
   verificationTokenExpires: Date,
   passwordResetToken: String,
-  passwordResetExpires: Date
+  passwordResetExpires: Date,
+  referralCode: { type: String, unique: true },
 }, {
   timestamps: true
 });
