+++ backend/routes/authRoutes.js
@@ -13,7 +13,7 @@ router.post('/college-admin', authController.login_college);
 
 router.post('/company-admin', authController.login_company);
 
-router.post('/register/check-otp', authController.generateRegistrationOtp);
+router.post('/register/check-otp', authController.checkRegistrationOtp);
 router.post('/logout', authController.logout_college_company);
 
 // Add this test route
